\name{CCMO-package}
\alias{CCMO-package}
\docType{package}
\title{
A powerful statistical method for disease-SNP association analysis using case-control mother-offspring pair data.
}
\description{
The statistical method implemented in this package is designed to detect SNPs associated with offspring disease using the genotype data from mother-offspring pairs by incorporting maternal covariates. The method allows for main SNP effects of mother and offspring, the interaction between maternal and offspring genotypes, and the interactions between the SNP genotypes and maternal covariates. The disease prevalence formula should be specified by users. In this method, Hardy-Weiberg equilibrium can be incorprated and mode of inheritance (recessive, additive, and dominant) and be specified. The maternal genotype can be specified to be either dependent (through the so called daLOG model) or independent of the maternal covariates.
}
\references{
Zhang H, Mukherjee B, Arthur V, Hu G, Hochner H, Chen J (2019). An Efficient and Computationally Robust Statistical Method for Analyzing Case-Control Mother-Offspring Pair Genetic Association Studies. Accepted by Annals of Applied Statistics.
}

\author{
Hong Zhang, Jinbo Chen

Maintainer: Hong Zhang <zhangh@ustc.edu.cn>
}
\keyword{ package }
