% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCMO.R
\name{MultiLociPOE.input}
\alias{MultiLociPOE.input}
\title{Inputs for the function MultiLociPOE}
\usage{
MultiLociPOE.input(gmm, gcc, type)
}
\arguments{
\item{gmm}{a \code{n} x \code{m} matrix of genotypes for mothers (n: number of mothers; m: number of SNPs). The possible values should be 0, 1, 2.}

\item{gcc}{a \code{n} x \code{m} matrix of genotypes for offspring (n: number of offspring; m: number of SNPs). The possible values should be 0, 1, 2.}

\item{type}{an indicator for the genotype data format. (0 for genotype coded as 0, 1, 2; 1 for normalized genotype coded as 0, 1/2, 1; 2 for genotype coded like "A/A", "A/B", "B/B")}
}
\value{
a list with the following elements

\item{\code{gmm}}{matrix of genotypes for mothers (n: number of mothers; m: number of SNPs)}

\item{\code{gcc}}{matrix of genotypes for offspring (n: number of offspring; m: number of SNPs)}

\item{\code{hap}}{a \code{l} x \code{m} matrix of possible haplotypes in the cohort (l: number of possible haplotypes; m: number of SNPs)}

\item{\code{ppi}}{a \code{l}-vector of the corresponding haplotype frequencies (l: number of possible haplotypes)}
}
\description{
This function can be used to obtain required inputs for the function MultiLociPOE: all possible haplotypes in the population of interest and the corresponding haplotype frequencies.
}
