% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCMO.R
\name{MultiLociPOE}
\alias{MultiLociPOE}
\title{POEs analysis using multi-locus genotype data from case-control mother-offspring pairs}
\usage{
MultiLociPOE(Y, gmm, gcc, X, loci, hap, f, ppi)
}
\arguments{
\item{Y}{a \code{n}-vector of disease statuses for \code{n} offspring (1 for case and 0 for control).}

\item{gmm}{a \code{n} x \code{m} matrix of genotypes for mothers (\code{n}: number of mothers; \code{m}: number of SNPs). The possible values should be 0, 1, 2.}

\item{gcc}{a \code{n} x \code{m} matrix of genotypes for offspring (\code{n}: number of offspring; \code{m}: number of SNPs). The possible values should be 0, 1, 2.}

\item{X}{a \code{n}-vector of maternal covariates for main effects (\code{n}: number of mothers).}

\item{loci}{an indicator for the test locus. The possible values could be 1,...,\code{m} (\code{m}: number of SNPs).}

\item{hap}{a \code{l} x \code{m} matrix of possible haplotypes in the population of interest (\code{l}: number of possible haplotypes; \code{m}: number of SNPs).}

\item{f}{specified disease prevalence.}

\item{ppi}{a \code{l}-vector of the corresponding haplotype frequencies (\code{m}: number of possible haplotypes).}
}
\value{
a list with the following elements

\item{\code{new}}{estimation and significance test results for the new method M-HAP}

\item{\code{log}}{estimation and significance test results for the standard logistic regression method}

\item{\code{cov.new}}{covariance matrix of the estimated parameters by the new method M-HAP}

\item{\code{cov.log}}{covariance matrix of the estimated parameters by the standard logistic regression method)}
}
\description{
This function implements the statistical method M-HAP for detecting POEs using genotype data from case-control mother-offspring pairs by adjusting for covariates. M-HAP utilizes available information such as Medelian inheritance law, Hardy-Weinberg equilibrium, and conditional independence between offspring genotypes and maternal covariates given maternal genotypes.
}
\examples{
\dontrun{
data(POESampleData)
Y = POESampleData[,1]
gmm = POESampleData[,2:6]
gcc = POESampleData[,7:11]
X = POESampleData[,12]
loci = 1
f = 0.01
data = MultiLociPOE.input(gmm,gcc,0)
gmm = data$gmm
gcc = data$gcc
hap = data$hap
ppi = data$ppi
fit = MultiLociPOE(Y,gmm,gcc,X,loci,hap,f,ppi)
}
}
