% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCMO.R
\name{multipleSNP}
\alias{multipleSNP}
\title{Association analysis with multiple SNPs}
\usage{
multipleSNP(
  Y,
  Gc,
  Gm,
  Xo = NULL,
  Xc = NULL,
  Xm = NULL,
  X.Gm = NULL,
  G.main = c("Gc", "Gm"),
  G.int = FALSE,
  mode = "add",
  prev = 0.01,
  ind = FALSE,
  HWE = TRUE,
  normalized.genotype = FALSE,
  test = NULL,
  cores = NULL
)
}
\arguments{
\item{Y}{a \code{n}-vector of disease statuses for n offspring (1 for case and 0 for control).}

\item{Gc}{a \code{n} x \code{m} matrix of genotypes for offspring (n: number of offspring; m: number of SNPs). The possible values should be 0, 1, 2, or 1/2 (see \code{normalized.genotype} for details).}

\item{Gm}{a \code{n} x \code{m} matrix of genotypes for mothers (n: number of mothers; m: number of SNPs). The possible values should be 0, 1, 2, or 1/2 (see \code{normalized.genotype} for details).}

\item{Xo}{a \code{n} x \code{n_o} matrix of maternal covariates for main effects (n: number of mothers; \code{n_o}: the numbers of genotypes; default value: NULL)}

\item{Xc}{a \code{n} x \code{n_c} matrix of maternal covariates for intercation effects with offspring genotypes (\code{n}: number of mothers; \code{n_c}: the numbers of genotypes; default value: NULL)}

\item{Xm}{a \code{n} x \code{n_m} matrix of maternal covariates for intercation effects with maternal genotypes (\code{n}: number of mothers; \code{n_m}: the numbers of genotypes; default value: NULL)}

\item{X.Gm}{a \code{n} x \code{n_g} matrix of maternal covariates potentially associated with maternal genotypes (\code{n}: number of mothers; \code{n_g}: the numbers of genotypes; default value: NULL)}

\item{G.main}{a vector containing the main SNP effects taking values 'Gc' and/or 'Gm' (default value: c('Gc','Gm'))}

\item{G.int}{an indicator for the presence of Gc x Gm interaction (TRUE for yes and FALSE for no; default value: FALSE)}

\item{mode}{mode of inheritance ('rec' for recessive, 'add' for additive, 'dom' for dominant; default value: 'add')}

\item{prev}{specified disease prevalence (default value: 0.01)}

\item{ind}{a logical variable indicating whether Gm and X.Gm are independent (TRUE for independence and FALSE for dependence, default value: FALSE)}

\item{HWE}{a logical variable indicating whether the HWE assumption is incorporated (TRUE for incorprated and FALSE otherwise; default value: TRUE)}

\item{normalized.genotype}{a logical variable indicating whether the genotypes are normalized (TRUE for normalized and FALSE otherwise, default value: FALSE). If FALSE, Gc and Gm should take values 0 (genotype \code{AA}), 1 (genotype \code{AB}), 2 (genotype \code{BB}); otherwise, their values depend on the mode of inheritance ('add': 0 for genotype \code{AA}, 0.5 for genotype \code{AB}, 1 for genotype \code{BB}; 'rec': 0 for genotype \code{AA} or AB, 1 for genotype BB; 'dom': 0 for genotype \code{AA}, 1 for genotype AB or BB).}

\item{test}{a vector of predictor indices in Omnibus test (default value: NULL; if not NULL, call \link{OmnibusTest})}

\item{cores}{the number of CPU cores used for parallele execution. If not specified, the system will determine its value.}
}
\value{
a list of length \code{m}, each element is a list with the following elements

\item{\code{new}}{estimation and significance test results for the new method}

\item{\code{log}}{estimation and significance test results for the standard logistic regression method}

\item{\code{cov.new}}{covariance matrix of the estimated parameters by the new method}

\item{\code{cov.log}}{covariance matrix of the estimated parameters by the standard logistic regression method)}

\item{\code{penetrance}}{logistic regression model for the penetrance function}

\item{\code{daLOG}}{daLOG model relating maternal genotype and maternal covariates}

\item{\code{Omnibus}}{Omnibus test results (test statistic, degrees of freedom, p-value) if \code{test} is not NULL}
}
\description{
Multiple SNPs can be analyzed utilizing multiple CPU cores (requires the R package \link{doMC}), which calls \link{singleSNP}. Estimation results include the parameter estimates and their estimated standard errors, the p-values for significance tests (need to call \link{OmnibusTest}), some model information, and omnibus test if required (the argument \code{test} should be specified).
}
\examples{
\dontrun{
library(foreach)
data(SampleData)
Y = SampleData[,1]
M = 10 # the number of SNPs
Gc = SampleData[,1+1:M]
Gm = SampleData[,1+M+(1:M)]
X = SampleData[,-(1:(1+2*M))]
res = multipleSNP(Y,Gc,Gm,Xo=X,Xc=X,Xm=X,X.Gm=X)
}
}
\seealso{
\code{\link{singleSNP}} \code{\link{OmnibusTest}}
}
