% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCMO.R
\name{OmnibusTest}
\alias{OmnibusTest}
\title{Omnibus test for multiple effects}
\usage{
OmnibusTest(d, test)
}
\arguments{
\item{d}{an object returned by the R function \link{singleSNP}}

\item{test}{a vector of predictor indices in Omnibus test (default value: NULL)}
}
\value{
a list with an additional element for Omnibus test p-values

\item{\code{Omnibus}}{Omnibus test results (test statistic, degrees of freedom, p-value) if \code{test} is not NULL}
}
\description{
The interested effects include the genetic effects and/or environment effects and/or gene-environment interaction effects and/or gene-gene interaction effects. Association analysis is conducted between a single SNP and the disease of interest.
}
\examples{
\dontrun{
data(SampleData)
Y = SampleData[,1]
M = 10 # the number of SNPs
Gc = SampleData[,2]
Gm = SampleData[,12]
X = SampleData[,-(1:(1+2*M))]
d = singleSNP(Y,Gc,Gm,Xo=X,Xc=X,Xm=X,X.Gm=X)
test = c(3,4,6,7) # main SNP effects and SNP-covariate interaction effects
d = OmnibusTest(d,test)
}
}
\seealso{
\code{\link{singleSNP}} \code{\link{multipleSNP}}
}
