% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCMO.R
\name{SampleData}
\alias{SampleData}
\title{A sample dataset}
\description{
This sample dataset include the genotypes for 2000 mother-offspring pairs (1000 with case offspring and 1000 with control offspring), together with two covariate values for each of the 2000 mothers.
}
\examples{
\dontrun{
data(SampleData)
}
}
