% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCMO.R
\name{singleSNP}
\alias{singleSNP}
\title{Association analysis with a single SNP}
\usage{
singleSNP(
  Y,
  Gc,
  Gm,
  Xo = NULL,
  Xc = NULL,
  Xm = NULL,
  X.Gm = NULL,
  G.main = c("Gc", "Gm"),
  G.int = FALSE,
  mode = "add",
  prev = 0.01,
  ind = FALSE,
  HWE = TRUE,
  normalized.genotype = FALSE
)
}
\arguments{
\item{Y}{a \code{n}-vector of disease statuses for n offspring (1 for case and 0 for control)}

\item{Gc}{a \code{n}-vector of genotypes for offspring, taking values 0, 1, and 2 (n: number of offspring).  The possible values should be 0, 1, 2, or 1/2 (see \code{normalized.genotype} for details).}

\item{Gm}{a \code{n}-vector of genotypes for mothers, taking values 0, 1, and 2 (n: number of mothers).  The possible values should be 0, 1, 2, or 1/2 (see \code{normalized.genotype} for details).}

\item{Xo}{a \code{n} x \code{n_o} matrix of maternal covariates for main effects (n: number of mothers; \code{n_o}: the numbers of genotypes; default value: NULL)}

\item{Xc}{a \code{n} x \code{n_c} matrix of maternal covariates for intercation effects with offspring genotypes (\code{n}: number of mothers; \code{n_c}: the numbers of genotypes; default value: NULL)}

\item{Xm}{a \code{n} x \code{n_m} matrix of maternal covariates for intercation effects with maternal genotypes (\code{n}: number of mothers; \code{n_m}: the numbers of genotypes; default value: NULL)}

\item{X.Gm}{a \code{n} x \code{n_g} matrix of maternal covariates potentially associated with maternal genotypes (\code{n}: number of mothers; \code{n_g}: the numbers of genotypes; default value: NULL)}

\item{G.main}{a vector containing the main SNP effects taking values 'Gc' and/or 'Gm' (default value: c('Gc','Gm'))}

\item{G.int}{an indicator for the presence of Gc x Gm interaction (TRUE for yes and FALSE for no; default value: FALSE)}

\item{mode}{mode of inheritance ('rec' for recessive, 'add' for additive, 'dom' for dominant; default value: 'add')}

\item{prev}{specified disease prevalence (default value: 0.01)}

\item{ind}{a logical variable indicating whether Gm and X.Gm are independent (TRUE for independence and FALSE for dependence, default value: FALSE)}

\item{HWE}{a logical variable indicating whether the HWE assumption is incorporated (TRUE for incorprated and FALSE otherwise; default value: TRUE)}

\item{normalized.genotype}{a logical variable indicating whether the genotypes are normalized (TRUE for normalized and FALSE otherwise, default value: FALSE). If FALSE, Gc and Gm should take values 0 (genotype \code{AA}), 1 (genotype \code{AB}), 2 (genotype \code{BB}); otherwise, their values depend on the mode of inheritance ('add': 0 for genotype \code{AA}, 0.5 for genotype \code{AB}, 1 for genotype \code{BB}; 'rec': 0 for genotype \code{AA} or AB, 1 for genotype BB; 'dom': 0 for genotype \code{AA}, 1 for genotype AB or BB).}
}
\value{
a list with the following elements

\item{\code{new}}{estimation and significance test results for the new method}

\item{\code{log}}{estimation and significance test results for the standard logistic regression method}

\item{\code{cov.new}}{covariance matrix of the estimated parameters by the new method}

\item{\code{cov.log}}{covariance matrix of the estimated parameters by the standard logistic regression method)}

\item{\code{penetrance}}{logistic regression model for the penetrance function}

\item{\code{daLOG}}{daLOG model relating maternal genotype and maternal covariates}
}
\description{
Association analysis is conducted between a single SNP and the disease of interest. Estimation results include the parameter estimates and their estimated standard errors, the p-values for significance tests, covariance matrices, and some model information.
}
\examples{
\dontrun{
data(SampleData)
Y = SampleData[,1]
M = 10 # the number of SNPs
Gc = SampleData[,2]
Gm = SampleData[,12]
X = SampleData[,-(1:(1+2*M))]
res = singleSNP(Y,Gc,Gm,Xo=X,Xc=X,Xm=X,X.Gm=X)
}
}
\seealso{
\code{\link{multipleSNP}} \code{\link{OmnibusTest}}
}
